+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2019-02-15
+### Added
+
 ## [0.3.3] - 2019-02-15
 ### Added
 - `android_location.Location` class stub
@@ -173,6 +176,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `android_content.Context.packageName` getter
 - `android_content.Context.packageResourcePath` getter
 
+[0.3.4]:  https://github.com/drydart/flutter_android/compare/0.3.3...0.3.4
 [0.3.3]:  https://github.com/drydart/flutter_android/compare/0.3.2...0.3.3
 [0.3.2]:  https://github.com/drydart/flutter_android/compare/0.3.1...0.3.2
 [0.3.1]:  https://github.com/drydart/flutter_android/compare/0.3.0...0.3.1